% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{exportTree}
\alias{exportTree}
\title{Export a distance matrix as a tree object}
\usage{
exportTree(getColorDistanceMatrixObject, file, returnTree = F)
}
\arguments{
\item{getColorDistanceMatrixObject}{A distance matrix, especially as returned by \code{\link{getColorDistanceMatrix}}, but any numeric symmetrical matrix will work.}

\item{file}{Character vector of desired filename for saving tree. Should end in ".newick".}

\item{returnTree}{Logical. Should the tree object be returned to the working environment in addition to being saved as a file?}
}
\value{
Newick tree saved in specified location and \code{as.phylo} tree object if \code{returnTree=TRUE}.
}
\description{
Converts a symmetrical distance matrix to a tree and saves it in newick format. Uses \code{\link[stats]{hclust}} to form clusters.
}
\examples{
clusterList <- getHistList("Heliconius/")
CDM <- getColorDistanceMatrix(clusterList, method="emd", plotting=F)
heliconius_tree <- exportTree(CDM, "./HeliconiusColorTree.newick", returnTree=T)
}
