% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{imageClusterPipeline}
\alias{imageClusterPipeline}
\title{Generate and plot a color distance matrix from a set of images}
\usage{
imageClusterPipeline(images, clusterMethod = "hist", distanceMethod = "emd",
  lower = c(0, 140/255, 0), upper = c(60/255, 1, 60/255), histBins = 3,
  kmeansBins = 27, binAvg = T, normPix = F, plotBins = F, pausing = T,
  hsv = F, bounds = c(0, 1), sampleSize = 20000, iter.max = 50,
  nstart = 5, imgType = F, ordering = "default", sizeWeight = 0.5,
  colorWeight = 0.5, plotHeatmap = T, returnDistanceMatrix = T,
  saveTree = F, saveDistanceMatrix = F)
}
\arguments{
\item{images}{Character vector of directories, image paths, or both.}

\item{clusterMethod}{Which method for getting color clusters from each image
should be used? Must be either \code{"hist"} (predetermined bins generated
by dividing each channel with equidistant bounds; calls
\code{\link{getHistList}}) or \code{"kmeans"} (determine clusters using
kmeans fitting on pixels; calls \code{\link{getKMeansList}}).}

\item{distanceMethod}{One of four possible comparison methods for calculating
the color distances: \code{"emd"} (uses \code{\link{EMDistance}},
recommended), \code{"chisq"} (uses \code{\link{chisqDistance}}),
\code{"color.dist"} (uses \code{\link{colorDistance}}; not appropriate if
binAvg=F), or \code{"weighted.pairs"}
(\code{\link{weightedPairsDistance}}).}

\item{lower}{RGB or HSV triplet specifying the lower bounds for background
pixels. Default upper and lower bounds are set to values that work well for
a bright green background (RGB [0, 1, 0]).}

\item{upper}{RGB or HSV triplet specifying the upper bounds for background
pixels. Default upper and lower bounds are set to values that work well for
a bright green background (RGB [0, 1, 0]). Determining these bounds may
take some trial and error, but the following bounds may work for certain
common background colors: \itemize{ \item Black: lower=c(0, 0, 0);
upper=c(0.1, 0.1, 0.1) \item White: lower=c(0.8, 0.8, 0.8); upper=c(1, 1,
1) \item Green: lower=c(0, 0.55, 0); upper=c(0.24, 1, 0.24) \item Blue:
lower=c(0, 0, 0.55); upper=c(0.24, 0.24, 1) } If no background filtering is
needed, set bounds to some non-numeric value (\code{NULL}, \code{FALSE},
\code{"off"}, etc); any non-numeric value is interpreted as \code{NULL}.}

\item{histBins}{Only applicable if \code{clusterMethod="hist"}. Number of bins
for each channel OR a vector of length 3 with bins for each channel. Bins=3
will result in 3^3 = 27 bins; bins=c(2, 2, 3) will result in 2*2*3=12 bins
(2 red, 2 green, 3 blue), etc. Passed to \code{\link{getHistList}}.}

\item{kmeansBins}{Only applicable if \code{clusterMethod="kmeans"}. Number of
KMeans clusters to fit. Unlike \code{\link{getImageHist}}, this represents
the actual final number of bins, rather than the number of breaks in each
channel.}

\item{binAvg}{Logical. Should the color clusters used for the distance matrix
be the average of the pixels in that bin (binAvg=\code{TRUE}) or the center
of the bin ({FALSE})? If a bin is empty, the center of the bin is returned
as the cluster color regardless. Only applicable if
\code{clusterMethod="hist"}, since \code{kmeans} clusters are at the center
of their assigned pixel clouds by definition.}

\item{normPix}{Logical. Should RGB or HSV cluster values be normalized using
\code{\link{normalizeRGB}}?}

\item{bounds}{Upper and lower limits for the channels; R reads in images with
intensities on a 0-1 scale, but 0-255 is common.}

\item{sampleSize}{Only applicable if \code{clusterMethod="kmeans"}. Number of
pixels to be randomly sampled from filtered pixel array for performing fit.
If set to \code{FALSE}, all pixels are fit, but this can be time-consuming,
especially for large images. Passed to \code{\link{getKMeansList}}.}

\item{iter.max}{Only applicable if \code{clusterMethod="kmeans"}. Inherited
from \code{\link[stats]{kmeans}}. The maximum number of iterations allowed
during kmeans fitting. Passed to \code{\link{getKMeansList}}.}

\item{nstart}{Only applicable if \code{clusterMethod="kmeans"}. Inherited
from \code{\link[stats]{kmeans}}. How many random sets should be chosen?
Passed to \code{\link{getKMeansList}}.}

\item{ordering}{Logical if not left as "default". Should the color clusters
in the list be reordered to minimize the distances between the pairs? If
left as default, ordering depends on distance method: "emd" and "chisq" do
not order clusters ("emd" orders on a case-by-case in the
\code{\link{EMDistance}} function itself and reordering by size similarity
would make chi-squared meaningless); "color.dist" and "weighted.pairs" use
ordering. To override defaults, set to either \code{T} (for ordering) or
\code{F} (for no ordering).}

\item{sizeWeight}{Weight of size similarity in determining overall score and
ordering (if \code{ordering=T}).}

\item{colorWeight}{Weight of color similarity in determining overall score
and ordering (if \code{ordering=T}). Color and size weights do not
necessarily have to sum to 1.}

\item{plotHeatmap}{Logical. Should a heatmap of the distance matrix be
plotted?}

\item{returnDistanceMatrix}{Logical. Should the distance matrix be returned
to the R environment or just plotted?}

\item{saveTree}{Either logical or a filepath for saving the tree; default if
set to \code{TRUE} is to save in current working directory as
"ColorTree.newick".}

\item{saveDistanceMatrix}{Either logical or filepath for saving distance
matrix; default if set to \code{TRUE} is to save in current working
directory as "ColorDistanceMatrix.csv"}
}
\value{
Color distance matrix, heatmap, and saved distance matrix and tree
  files if saving is \code{TRUE}.
}
\description{
Takes images, computes color clusters for each image, and calculates distance
matrix/dendrogram from those clusters. This is the fastest way to get a
distance matrix for color similarity starting from a folder of images.
Essentially, it just calls in a series of other package functions in order:
input images -> \code{\link{getImagePaths}} -> \code{\link{getHistList}} or
\code{\link{getKMeansList}} followed by \code{\link{extractClusters}} ->
\code{\link{getColorDistanceMatrix}} -> plotting -> return/save distance
matrix. Sort of railroads you, but good for testing different combinations of
clustering methods and distance metrics.
}
\examples{
imageClusterPipeline("Heliconius/")
}
