% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_miscellaneous.R
\name{pause}
\alias{pause}
\title{Pause and wait for user input}
\usage{
pause()
}
\description{
Tiny little function wrapper, mostly used for looping or when several plots
are output by a single function. Waits for user keystroke to move on to next
image or exit.
}
\examples{
for (i in c(1:5)) {
  print(i)
  if (i < 5) {
    colordistance:::pause()
  }
}
}
